tests: refactor Telethon fixtures and remove redundant imports
- Added reusable fixtures for `telegram_chat_id`, `telegram_user_id`, and `telegram_chat_participants` in `telethon.py`.
- Modified test services to use updated fixtures.
- Removed duplicate and unused imports in `test_supertelethon.py`.
